// src/components/select/select-input.component.d.ts
import { EventEmitter, AfterViewInit } from '@angular/core';
export declare class SelectInputComponent implements AfterViewInit {
    placeholder: string;
    autofocus: boolean;
    allowClear: boolean;
    multiple: boolean;
    tagging: boolean;
    identifier: any;
    options: any[];
    label: string;
    allowAdditions: boolean;
    selected: any[];
    toggle: EventEmitter<any>;
    selection: EventEmitter<any>;
    activate: EventEmitter<any>;
    keyup: EventEmitter<any>;
    inputElement: any;
    readonly caretVisible: boolean;
    selectedOptions: any[];
    _selected: any[];
    ngAfterViewInit(): void;
    onKeyUp(event: any): void;
    onKeyDown(event: any): void;
    onClick(event: any): void;
    onOptionRemove(event: any, option: any): void;
    calcSelectedOptions(selected: any[]): any[];
}